pyproject.toml
[tool.poetry]
name = "viva_vdm"
version = "0.1.0"
description = "The Vaccine Design Module of the ViVA Project"
authors = ["Shan <stwm2@student.london.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.8.2"
celery = "^5.1.2"
uvicorn = "^0.15.0"
fastapi = "^0.68.1"
mongoengine = "^0.23.1"
python-dotenv = "^0.19.0"
backoff = "^1.11.1"
biopython = "^1.79"
mongoengine-goodjson = "^1.1.8"
mhcflurry = "^2.0.2"
fastapi-cache2 = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.14.1"

[tool.poetry.scripts]
setup = 'setup:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py38']
include = '''
/(
    \.pyi?$
  | \.pyt?$
)/
'''
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''